linalg/ldlt : API 's vectorD member func now returns Eigen::Map
+ vectorD initially returned Diagonal<> expression
+ now return view to underlying data
+ makes backend compatible with proxsuite's ldlt

+++ include/proxnlp/linalg/blocks.hxx
@@ -36,7 +36,7 @@ BlockLDLT<Scalar>::updateBlockPermutationMatrix(const SymbolicBlockMatrix &in) {
   using IndicesType = PermutationType::IndicesType;
   IndicesType &indices = m_permutation.indices();
   isize idx = 0;
-  for (isize i = 0; i < nblocks(); ++i) {
+  for (isize i = 0; i < (isize)nblocks(); ++i) {
     m_idx[(usize)i] = idx;
     idx += row_segs[i];
   }
